added auth

+++ src/resolvers.ts
@@ -2,13 +2,15 @@ import { PubSub } from "graphql-subscriptions";
 import {scrape} from './scraper';
 import {getDirectories, getFilenames, removeFile} from "./utils";
 import {SCREENSHOTS_PATH} from "./server";
-import {deleteJobByUuid} from "./mongodb";
+import {deleteJobByUuid, User} from "./mongodb";
+import {login, register} from "./auth";
 
 export const pubsub = new PubSub();
 
 export default {
   Query: {
-    getGroupNames: async () => {
+    getGroupNames: async (_: any, args: {}, context: { user: User }) => {
+      if (!context.user) throw authError('Unauthorized');
       try {
         return { names: getDirectories(SCREENSHOTS_PATH) };
       } catch (error) {
@@ -17,9 +19,18 @@ export default {
     },
     getScreenshotsByGroup: async (_: any, args: {
       groupName: string,
-    }, __: any) => {
+    }, context: { user: User }) => {
+      if (!context.user) throw authError('Unauthorized');
       try {
         return { files: getFilenames(`${SCREENSHOTS_PATH}/${args.groupName}`) };
+      } catch (error) {
+        throw error;
+      }
+    },
+    verify: async (_: any, args: {}, context: { user: User }) => {
+      if (!context.user) throw authError('Unauthorized');
+      try {
+        return { user: context.user };
       } catch (error) {
         throw error;
       }
@@ -32,7 +43,9 @@ export default {
       jobAnchorSelector: string,
       jobLinkContains: string,
       numberOfPages: number,
-      }, __: any) => {
+      }, context: { user: User }) => {
+      if (!context.user) throw authError('Unauthorized');
+
       const _host = args.host.replace(/&quot/g, '"');
       const _path = args.path.replace(/&quot/g, '"');
       const _jobAnchorSelector = args.jobAnchorSelector.replace(/&quot/g, '"');
@@ -42,7 +55,8 @@ export default {
     removeScreenshotByGroupAndUuid: async (_: any, args: {
       groupName: string,
       uuid: string
-    }, __: any) => {
+    }, context: { user: User }) => {
+      if (!context.user) throw authError('Unauthorized');
       try {
         const scrRemoved = removeFile(`${SCREENSHOTS_PATH}/${args.groupName}/_${args.uuid}.png`);
         const jobRemoved = deleteJobByUuid(args.uuid);
@@ -50,11 +64,65 @@ export default {
       } catch (error) {
         throw error;
       }
+    },
+    login: async (_: any, args: {
+      email: string,
+      password: string
+    }, __: any) => {
+      try {
+        const { success, error, token, user } = await login(args.email, args.password);
+        if (success && token && user) {
+          return { success, code: 200, token, user };
+        } else if (error) {
+          throw clientError(error.message);
+        }
+      } catch (error) {
+        throw serverError(error.message);
+      }
+    },
+    register: async (_: any, args: {
+      email: string,
+      password: string
+    }, __: any) => {
+      try {
+        const { success, error, token, user } =  await register(args.email, args.password);
+        if (success && token && user) {
+          return { success, code: 200, token, user };
+        } else if (error) {
+          throw clientError(error.message);
+        }
+      } catch (error) {
+        throw serverError(error.message);
+      }
     }
   },
   Subscription: {
     newJobs: {
-      subscribe: () => pubsub.asyncIterator(['newJobs'])
+      subscribe: (_: any, args: {}, context: { user: User }) => {
+        if (!context.user) throw authError('Unauthorized');
+        return pubsub.asyncIterator(['newJobs'])
+      }
     }
   }
 };
+
+const clientError = (msg: string) =>
+    new Error(JSON.stringify({
+      success: false,
+      code: 400,
+      msg
+    }));
+
+const serverError = (msg: string) =>
+    new Error(JSON.stringify({
+      success: false,
+      code: 500,
+      msg
+    }));
+
+const authError = (msg: string) =>
+    new Error(JSON.stringify({
+      success: false,
+      code: 403,
+      msg
+    }));
